Refactoring after code review with Olivier
